abstractionLogic
