+++ src/Apropos/HasPermutationGenerator/Abstraction.hs
@@ -4,13 +4,14 @@ module Apropos.HasPermutationGenerator.Abstraction (
   Abstraction (..),
   abstract,
   gotoSum,
+  abstractLogic,
   abstractsProperties,
 ) where
 
 import Apropos.HasParameterisedGenerator
 import Apropos.HasPermutationGenerator.Contract
 import Apropos.HasPermutationGenerator.Morphism
-import Apropos.LogicalModel (Formula (..), satisfiedBy)
+import Apropos.LogicalModel (Formula (..), LogicalModel(logic), satisfiedBy)
 import Apropos.LogicalModel.Enumerable
 import Control.Lens
 import Data.Either (rights)
@@ -28,8 +29,8 @@ data Abstraction ap am bp bm
       , propertyAbstraction :: Prism' bp ap
       , sumModelAbstraction :: Prism' bm am
       , propLabel :: bp
-      , modelConstructor :: am -> bm
-      , propConstructor :: ap -> bp
+      --, modelConstructor :: am -> bm
+      --, propConstructor :: ap -> bp
       }
 
 abstract ::
@@ -73,10 +74,10 @@ gotoSum s@SumAbstraction {} =
     Morphism
       { name = "goto " <> abstractionName s
       , match = Not $ Var $ propLabel s
-      , contract = clear >> add (propLabel s) >> addAll (propConstructor s <$> satisfiedBy)
+      , contract = clear >> add (propLabel s) >> addAll ((propertyAbstraction s #) <$> satisfiedBy)
       , morphism =
           const $
-            modelConstructor s
+            (sumModelAbstraction s  #)
               <$> genSatisfying @ap @am (All $ Var <$> satisfiedBy)
       }
 
@@ -107,3 +108,10 @@ abstractContract prefix a c = do
     projectProperties pa s = Set.fromList $ rights (matching pa <$> Set.toList s)
     maskProperties :: Prism' b a -> Set b -> Set b
     maskProperties pa = Set.filter (isn't pa)
+
+abstractLogic :: forall bp bm ap am. LogicalModel ap => Abstraction ap am bp bm -> Formula bp
+abstractLogic s@ProductAbstraction{} = (propertyAbstraction s #) <$> logic
+abstractLogic s@SumAbstraction{} =
+  (Var (propLabel s) :->: (propertyAbstraction s #) <$> logic)
+  :&&: (Not (Var (propLabel s)) :->: None (Var . (propertyAbstraction s #) <$> enumerated))
+
